requirements-docs.txt
ipython
numpydoc
pydocstyle
Sphinx>=1.8.3
pydata-sphinx-theme>=0.13.0
sphinx-panels
sphinx-notfound-page
sphinx-copybutton
sphinxcontrib-mermaid
sphinxcontrib-jsmath
nbsphinx
sphinx-gallery
myst_parser
tensorboard